Added game route
