+++ src/api/index.js
@@ -7,97 +7,15 @@ export {
   FETCH_CHARACTERS_QUERY,
   UPDATE_CHARACTER_MUTATION
 } from './characters'
-
-export const FETCH_GAMES_QUERY = gql`
-  {
-    games {
-      id
-      title
-      image
-      description
-      startingDate
-      lvlFrom
-      lvlTo
-      players
-      status
-      characters {
-        id
-        name
-        experience
-        renown
-        faction {
-          id
-          name
-          logo
-        }
-        user {
-          id
-          firstName
-          lastName
-          avatar
-          username
-        }
-      }
-      user {
-        id
-        firstName
-        lastName
-        avatar
-        username
-      }
-    }
-  }
-`
-
-export const CREATE_GAME_QUERY = gql`
-  mutation CreateGame(
-  $title: String!
-  $image: String!
-  $description: String!
-  $startingDate: String!
-  $lvlFrom: Int!,
-  $lvlTo: Int!,
-  $players: Int!,
-  $telegramPost: Boolean!,
-  $facebookPost: Boolean!,
-  $gameForNewbies: Boolean!,
-  $isAl: Boolean!,
-  ){
-    createGame(
-      title: $title
-      image: $image
-      description: $description
-      startingDate: $startingDate
-      lvlFrom: $lvlFrom
-      lvlTo: $lvlTo
-      players: $players,
-      telegramPost: $telegramPost,
-      facebookPost: $facebookPost,
-      gameForNewbies: $gameForNewbies,
-      isAl: $isAl,
-    ) {
-      id
-      title
-      image
-      description
-      startingDate
-      lvlFrom
-      lvlTo
-      players
-      characters {
-        id,
-        name,
-        experience,
-        renown
-        faction {
-          id
-          name
-          logo
-        }
-      },
-    }
-  }
-`
+export {
+  CREATE_GAME_QUERY,
+  END_GAME,
+  FETCH_GAMES_QUERY,
+  FETCH_GAME_QUERY,
+  LEAVE_GAME,
+  NEW_GAME_SUBSCRIPTION,
+  PARTICIPATE_GAME,
+} from './games'
 
 export const FETCH_USERS_QUERY = gql`
   query Users($username: String!) {
@@ -151,40 +69,6 @@ export const UPDATE_USER_ROLES = gql`
   }
 `
 
-export const NEW_GAME_SUBSCRIPTION = gql`
-  subscription {
-    newGame {
-      id
-      title
-      image
-      description
-      startingDate
-      lvlFrom
-      lvlTo
-      players
-      characters {
-        id
-        name
-        experience
-        renown
-        faction {
-          id
-          name
-          logo
-        }
-      }
-      status
-      user {
-        id
-        firstName
-        lastName
-        avatar
-        username
-      }
-    }
-  }
-`
-
 export const SIGN_IN_MUTATION = gql`
   mutation SignIn(
     $id: ID!
@@ -204,62 +88,5 @@ export const SIGN_IN_MUTATION = gql`
       avatar: $avatar
       authDate: $authDate
     )
-  }
-`
-
-export const PARTICIPATE_GAME = gql`
-  mutation SignIn(
-    $gameId: ID!
-    $characterId: ID!
-  ) {
-    participateGame(
-      gameId: $gameId
-      characterId: $characterId
-    ) {
-      id
-      players
-      status
-      characters {
-        id
-        name
-        class
-        avatar
-        experience
-        renown
-        faction {
-          id
-          name
-          logo
-        }
-        user {
-          id
-          firstName
-          lastName
-          avatar
-          username
-        }
-      }
-    }
-  }
-`
-
-export const LEAVE_GAME = gql`
-  mutation LeaveGame(
-    $characterId: ID!
-  ){
-    leaveGame(characterId: $characterId){
-      id,
-      game {
-        id
-      }
-    }
-  }
-`
-
-export const END_GAME = gql`
-  mutation EndGame(
-    $gameId: ID!
-  ){
-    endGame(gameId: $gameId)
   }
 `
