+++ src/components/Game/index.js
@@ -1,0 +1,3 @@
+export { GameInfo } from './GameInfo'
+export { ParticipantsList } from './ParticipantsList'
+export { GameActions } from './GameActions'
