+++ src/containers/Calendar/index.js
@@ -5,7 +5,8 @@ import Calendar from './Calendar'
 const Scene = ({ match }) => {
   return (
     <Switch>
-      <Route exact path={match.path} component={Calendar}/>
+      <Route exact path={match.path} component={Calendar} />
+      <Route path={`${match.path}/:gameId`} component={Calendar} />
     </Switch>
   )
 }
